completed types for game
