+++ backend/model/types/board.go
@@ -1,0 +1,6 @@
+package types
+
+type Board struct {
+	Size  BoardSize
+	board [][]Piece
+}
