+++ backend/model/types/boardsize.go
@@ -1,0 +1,9 @@
+package types
+
+type BoardSize uint
+
+const (
+	Undefined BoardSize = 0
+	Full      BoardSize = 19
+	// TODO: the rest
+)
