+++ backend/model/types/colorchoice.go
@@ -1,0 +1,8 @@
+package types
+
+type ColorChoice string
+
+const (
+	White ColorChoice = "white"
+	Black ColorChoice = "black"
+)
