+++ backend/model/types/move.go
@@ -1,0 +1,15 @@
+package types
+
+type Coord struct {
+	x uint
+	y uint
+}
+
+type Move struct {
+	Piece *Piece // nil counts as Pass
+	Coord *Coord // nil when Piece is nil
+}
+
+func (m Move) IsPass() bool {
+	return m.Piece == nil
+}
