+++ backend/model/types/piece.go
@@ -1,4 +1,4 @@
-package model
+package types
 
 type Piece int
